Fix bugs for builds under macOS and recent Eigen versions (Eigen > 3.3)

+++ include/STL_INC.hpp
@@ -16,6 +16,7 @@
 #include <chrono>
 #include <memory>
 #include <algorithm>
+#include <numeric>
 #include <utility>
 #include <limits>
 #ifdef WINDOWS
@@ -28,8 +29,8 @@
 
 
 
-typedef int64_t Integer;
-typedef uint64_t UInteger;
+typedef long long Integer;
+typedef unsigned long long UInteger;
 typedef double Real;
 typedef bool Boolean;
